Store envmaps in RGBE8 format instead of raw RGBA16F with a zero alpha channel.
This compresses slightly better (about 20% storage savings) and shouldn't
meaningfully affect the quality of the renders.

The changes are:

 - update device environment logic to process RGBE8 input instead of RGBA16F
 - handle f16 overflow properly during the conversion from RGBE8
 - invalidate the front-end asset cache by bumping the version
 - reprocess all the current HDRs (and add some new ones, including some small 2K options)

We should in theory be able to support 16K HDRs but it's untested and the PDF might run into
precision problems since it's stored in an F16 float in the renderer.